+++ cli/cmd/project_settings.py
@@ -1,6 +1,6 @@
 # Solution settings
 ENGINE_NAME = "richard" # Workspace name in premake file
-PROJECT_NAME = "application"  # Project name in premake file
+PROJECT_NAME = "client"  # Project name in premake file
 CONFIG = "debug" # Project config
 
 # VS settings
