+++ engine/src/core/window.h
@@ -1,6 +1,42 @@
 #pragma once
 
 /*
+* window_event represents the OS events that the method HandleEvents() may receive.
+* It is used to indicate those functions that call HandleEvents() which event
+* was popped out and processed.
+*/
+enum window_event {
+	EVENT_DEFAULT = 0,
+	EVENT_QUIT = 1
+};
+
+/*
+* window_initialize_response represents all the possible Initialize() scenarios.
+* It is possible that Initialize() succeeds or that fails
+* due to some internal initialization failing.
+*/
+enum window_initialize_response {
+	W_INTIALIZE_OK = 0,
+	W_INTIALIZE_SDL_WINDOW_FAIL = -1,
+	W_INITIALIZE_OPENGL_CONTEXT_FAIL = -2
+};
+
+/*
+* Constants realated to the window size
+*/
+const int WINDOW_WIDTH = 800;
+const int WINDOW_HEIGHT = 600;
+const int WINDOW_MIN_WIDTH = 400;
+const int WINDOW_MIN_HEIGHT = 200;
+
+/*
+* Constants related to the glad version downloaded
+*/
+const int GLAD_MAJOR_VERSION = 4;
+const int GLAD_MINOR_VERSION = 6;
+const int GLAD_DEPTH_SIZE = 24;
+
+/*
 * We forward declare this struct so every file that imports this header,
 * will not have to import anything SDL_Window related.
 * 
@@ -70,7 +106,7 @@ namespace Richard::Managers {
 		/*
 		* Pointer that points to the window that will receive the OS events.
 		*/
-		SDL_Window* mWindow;
+		SDL_Window* pWindow;
 
 		/*
 		* To use OpenGL within SDL, we need to capture the OpenGL Context
